Finished TrapBase and started spikes
