+++ Assets/Prefabs/Traps/TrapBase.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d2dde8bac47aaa841a71958adfdc26e0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
