+++ Assets/Scripts/Traps/Spikes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 57036681b6d3c6c43adef05f87bf55bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
